# api/doc/swagger_api_doc.json
{
  "info": {
    "title": "SHOPING API",
    "description": "This is a sample Petstore server. You can find \nout more about Swagger at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n",
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "thonguyen140294@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "swagger": "2.0",
  "host": "bestsporthome-api.herokuapp.com",
  "basePath": "/",
...
  "paths": {
...
    "/forgot-password/code": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "send code to reset password",
        "operationId": "sendCode",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "default": "en",
            "in": "header",
            "description": "Format language",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "phone munber",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Phone"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
...
    "/forgot-password/reset/{code}": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Reset password with code",
        "operationId": "changePassword",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "default": "en",
            "in": "header",
            "description": "Format language",
            "required": true,
            "type": "string"
          },
          {
            "name": "code",
            "in": "path",
            "description": "code reset password encoded",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "new password has encoded",
            "schema": {
              "$ref": "#/definitions/Reset_Password_Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
    "/user/kyc": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "update user kyc",
        "operationId": "kyc",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "default": "en",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Kyc_Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/user/change-password": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "update user password",
        "operationId": "changePassword",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "required": true,
            "type": "string",
            "default": "en"
          },
          {
            "in": "header",
            "name": "token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "password has encoded",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Change_Password_Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "get user profile",
        "operationId": "profile",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "default": "en",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Profile_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/user/cart": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "get user cart",
        "operationId": "cart",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "default": "en",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Cart_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "add cart",
        "operationId": "addCart",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "default": "en",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cart_Request"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "remove cart",
        "operationId": "removeCart",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "default": "en",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cart_Remove_Request"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "update cart qty",
        "operationId": "updateCart",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "locale",
            "default": "en",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "token",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cart_Update_Response"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/admin/view/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "get list user",
        "operationId": "getUsers",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "phone and password has encoded",
            "required": true,
            "type": "string"
          },
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/List_User_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
    "/admin/product": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "add product",
        "operationId": "addProduct",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "description": "phone and password has encoded",
            "required": true,
            "type": "string"
          },
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Add_Product_Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Message_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get list product",
        "operationId": "getProducts",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/List_Product_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
    "/product/{product_id}": {
      "get": {
        "tags": [
          "visitor"
        ],
        "summary": "get product detail",
        "operationId": "getProductDetail",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "locale",
            "in": "header",
            "default": "en",
            "description": "Select language return message",
            "required": true,
            "type": "string"
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Product_Detail_Response"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/Error_Response"
            }
          }
        }
      }
    },
...
  },
  "schemes": [
    "https",
    "http"
  ],
  "definitions": {
    "Message_Response": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "Error_Response": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },
...
    "Reset_Password_Request": {
      "type": "object",
      "properties": {
        "new_pass": {
          "type": "string"
        }
      },
      "required": [
        "new_pass"
      ]
    },
    "Change_Password_Request": {
      "type": "object",
      "properties": {
        "current_password": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        }
      }
    },
...
    "Profile_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "sex": {
          "type": "integer"
        },
        "birthday": {
          "type": "string"
        },
        "zip_code": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    },
    "List_User_Response": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer"
        },
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "docs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Profile_Response"
          }
        }
      }
    },
    "Phone": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string",
          "description": "phone number"
        }
      }
    },
    "Kyc_Request": {
      "type": "object",
      "properties": {
        "birthday": {
          "type": "string"
        },
        "sex": {
          "type": "integer"
        },
        "zip_code": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      }
    },
    "Product_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "category": {
          "type": "string"
        },
        "sale": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "Cart_Response": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string"
        },
        "product": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "image": {
              "type": "string"
            },
            "price": {
              "type": "integer"
            }
          }
        },
        "qty": {
          "type": "integer"
        },
        "status": {
          "type": "integer"
        }
      }
    },
    "Cart_Request": {
      "type": "object",
      "properties": {
        "product_id": {
          "type": "string"
        },
        "qty": {
          "type": "integer"
        }
      }
    },
    "Cart_Remove_Request": {
      "type": "object",
      "properties": {
        "cart_id": {
          "type": "string"
        }
      }
    },
    "Cart_Update_Response": {
      "type": "object",
      "properties": {
        "cart_id": {
          "type": "string"
        },
        "qty": {
          "type": "integer"
        }
      }
    },
    "Add_Product_Request": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "sale": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        }
      }
    },
...
    "List_Product_Response": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer"
        },
        "limit": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "docs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product_Response"
          }
        }
      }
    },
    "Detail_Response": {
      "type": "object",
      "properties": {
        "product": {
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "sizes": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Product_Detail_Response": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "category": {
          "type": "string"
        },
        "sale": {
          "type": "integer"
        },
        "image": {
          "type": "string"
        },
        "detail": {
          "$ref": "#/definitions/Detail_Response"
        }
      }
    },
...
  }
}